==> 1.HazardEstimation-AllTracksBbox.ipynb <==
## Description

This one 

from copy import deepcopy

import numpy as np
import pandas as pd
import cv2
import os
import os.path as osp

from tqdm import tqdm
from utils.ioutils import load_pickle
from utils.process_video import video_results_to_submission_format, process_video

%load_ext autoreload
%autoreload 2

ANNOTATIONS_PATH = "resources/annotations_public.pkl"
annotations = load_pickle(ANNOTATIONS_PATH)
VIDEOS_ROOT = "dataset/coool-benchmark"                       # <---- UPDATE THIS ONE

for video_name in annotations.keys():#Check that videos exist
    assert osp.exists(os.path.join(VIDEOS_ROOT, f"{video_name}.mp4"))

submission_results = {}

video_names = sorted(list(annotations.keys()))
for video_name in tqdm(video_names, total=len(video_names)):
    video_path = osp.join(VIDEOS_ROOT, f"{video_name}.mp4")
    video_results = process_video(video_path, annotations, run_tracks=True, run_captions=False)
    _submission_results = video_results_to_submission_format(video_results)
    submission_results.update(_submission_results)

df = pd.DataFrame.from_dict(submission_results, orient="index")
df.index.name = "ID"

remove_cols = ["Hazard_Track_ ", "Hazard_Name_ "]
df.drop(remove_cols, axis="columns", inplace=True, errors="ignore")
df

df.to_csv("submissions/results_bboxsizes_driver_alltracks.csv", sep=",")
